cycle_requirements.txt
absl-py==2.0.0
ansiwrap==0.8.4
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmapy==0.6.6
community==1.0.0b1
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
diffdist==0.1
fastjsonschema==2.18.0
filelock==3.12.3
find-libpython==0.3.0
Flask==3.0.3
fonttools==4.42.1
fsspec==2023.10.0
gdown==4.7.1
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
kmeans-pytorch==0.3
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.5
munkres==1.1.4
nbclient==0.8.0
nbformat==5.9.2
networkx==3.1
ninja==1.11.1
nlpaug==1.1.11
nnAudio==0.3.2
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
papermill==2.4.0
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-wavelets==1.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
sru==2.6.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
textwrap3==0.9.2
threadpoolctl==3.2.0
timm==0.4.5
torch==1.11.0
torchaudio==0.11.0
torchlibrosa==0.1.0
torchvision==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==3.0.1
wget==3.2
zipp==3.16.2
